requirements.txt
numpy>=2.0.0
pandas>=2.0.0
scikit-learn>=1.0.0
matplotlib>=3.7.1
Flask>=2.0.1
Flask-SocketIO>=5.3.6
Flask-SQLAlchemy>=2.5.1
Flask-Login>=0.5.0
Werkzeug>=2.0.1
python-socketio>=5.11.1
eventlet>=0.33.3
SQLAlchemy>=1.4.23
python-engineio>=4.8.2
itsdangerous>=2.1.2
Jinja2>=3.1.3
click>=8.1.7
bidict>=0.22.1
MarkupSafe>=2.1.5
six>=1.16.0
dnspython>=2.5.0
PyJWT>=2.8.0
flask-swagger-ui>=4.11.1
PyYAML>=6.0.1
python-dotenv>=0.19.0
fpdf>=1.7.2
twilio>=7.17.0
APScheduler>=3.8.1
seaborn>=0.13.2
XlsxWriter>=3.2.0
requests>=2.26.0
email-validator>=1.1.3
pytest>=6.2.5
black>=21.9b0
flake8>=3.9.2
